// types/user.ts
export interface IUserInfo {
	UserID: string
	Name: string
	NickName: string
	Grade: number
	Grade7: string
	selTopGrade: number
	selTopTerm: number
	Status: string
	StartPage: string
	MemberType: string
	LoginPath: string
	GradeUpChk17: string
	JoinRouteName: string
	SysUserId: string
	LoginSessionId: string
	LoginDateTime?: string
}